refactor: changed after review
