+++ packages/common/integration/tests/validators/date-interval-validator.spec.ts
@@ -1,10 +1,6 @@
-import { AbstractControl, FormControl, FormGroup, ValidatorFn } from '@angular/forms';
+import { FormControl, FormGroup } from '@angular/forms';
 import { toUtc } from '@angular-ru/common/date';
-import {
-    dateMaxIntervalValidator,
-    dateMinIntervalValidator,
-    orderedIntervalValidator
-} from '@angular-ru/common/validators';
+import { dateMaxIntervalValidator, dateMinIntervalValidator } from '@angular-ru/common/validators';
 
 describe('date interval validator', () => {
     it('max interval validator', () => {
@@ -60,48 +56,5 @@ describe('date interval validator', () => {
         // changing to invalid value again
         form.controls.dateFrom.setValue(toUtc({ month: new Date().getMonth() - 3 }));
         expect(form.errors).toEqual({ maxDateIntervalLimit: true });
-    });
-});
-
-describe('compare "from" and "to"', () => {
-    let form: FormGroup;
-    beforeEach(() => {
-        form = new FormGroup(
-            {
-                dateFrom: new FormControl(toUtc()),
-                dateTo: new FormControl(toUtc())
-            },
-            [orderedIntervalValidator({ dateToKey: 'dateTo', dateFromKey: 'dateFrom' })]
-        );
-    });
-
-    it('should not return error if "from = to"', () => {
-        expect(form.valid).toBe(true);
-    });
-
-    it('should not return error if "from < to"', () => {
-        form.controls.dateFrom.setValue(toUtc({ month: new Date().getMonth() - 1 }));
-        expect(form.valid).toBe(true);
-    });
-
-    it('should not return error if "from = null"', () => {
-        form.controls.dateFrom.setValue(null);
-        expect(form.valid).toBe(true);
-    });
-
-    it('should not return error if "to = null"', () => {
-        form.controls.dateFrom.setValue(null);
-        expect(form.valid).toBe(true);
-    });
-
-    it('should return error if "from > to"', () => {
-        form.controls.dateFrom.setValue(toUtc({ month: new Date().getMonth() + 1 }));
-        expect(form.errors).toEqual({ orderedInterval: true });
-    });
-
-    it('should return error if control is not part of FormGroup', () => {
-        const control: AbstractControl = new FormControl();
-        const validator: ValidatorFn = orderedIntervalValidator({ dateToKey: 'dateTo', dateFromKey: 'dateFrom' });
-        expect(() => validator(control).toThrow(new Error('DateIntervalValidator must be used on form group')));
     });
 });
